SAK-51816 SCORM Simplify status localization
