+++ scormplayer/scorm-tool/src/java/org/sakaiproject/scorm/ui/reporting/components/StatusLocalization.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright (c) 2007 The Apereo Foundation
+ *
+ * Licensed under the Educational Community License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *             http://opensource.org/licenses/ecl2
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.sakaiproject.scorm.ui.reporting.components;
+
+import java.util.Locale;
+
+import org.apache.commons.lang3.StringUtils;
+import org.apache.wicket.model.IModel;
+import org.apache.wicket.model.Model;
+import org.apache.wicket.model.ResourceModel;
+
+public final class StatusLocalization
+{
+	private StatusLocalization()
+	{
+	}
+
+	public static IModel<String> localizeStatus(String keyPrefix, String value)
+	{
+		if (StringUtils.isBlank(value))
+		{
+			return Model.of("");
+		}
+
+		String trimmedValue = value.trim();
+		String key = keyPrefix + trimmedValue.toLowerCase(Locale.ROOT).replace(' ', '_');
+		return new ResourceModel(key, trimmedValue);
+	}
+}
