+++ scormplayer/scorm-tool/src/java/org/sakaiproject/scorm/ui/reporting/pages/LearnerResultsPage.java
@@ -17,7 +17,6 @@ package org.sakaiproject.scorm.ui.reporting.pages;
 
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Locale;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.wicket.extensions.markup.html.repeater.data.table.IColumn;
@@ -41,7 +40,6 @@ import org.sakaiproject.wicket.ajax.markup.html.table.SakaiDataTable;
 import org.sakaiproject.wicket.markup.html.link.BookmarkablePageLabeledLink;
 import org.sakaiproject.wicket.markup.html.repeater.data.table.Action;
 import org.sakaiproject.wicket.markup.html.repeater.data.table.ActionColumn;
-import org.sakaiproject.wicket.markup.html.repeater.data.table.DecoratedPropertyColumn;
 
 public class LearnerResultsPage extends BaseResultsPage
 {
@@ -169,40 +167,5 @@ public class LearnerResultsPage extends BaseResultsPage
 		columns.add(new StatusPropertyColumn(successHeader, "successStatus", "successStatus", "success.status."));
 
 		return columns;
-	}
-
-	/**
-	 * Translates a raw SCORM status value (e.g. "completed", "passed") via the key "{keyPrefix}{value}",
-	 * falling back to the raw value when no translation exists.
-	 */
-	private class StatusPropertyColumn extends DecoratedPropertyColumn
-	{
-		private static final long serialVersionUID = 1L;
-
-		private final String keyPrefix;
-
-		public StatusPropertyColumn(IModel displayModel, String sortProperty, String propertyExpression, String keyPrefix)
-		{
-			super(displayModel, sortProperty, propertyExpression);
-			this.keyPrefix = keyPrefix;
-		}
-
-		@Override
-		public Object convertObject(Object object)
-		{
-			if (object == null)
-			{
-				return "";
-			}
-
-			String value = object.toString().trim();
-			if (value.isEmpty())
-			{
-				return "";
-			}
-
-			String key = keyPrefix + value.toLowerCase(Locale.ROOT).replace(' ', '_');
-			return getLocalizer().getString(key, LearnerResultsPage.this, value);
-		}
 	}
 }
